Refactor application structure and update documentation
- Reorganize code into modular package structure (grokparty/) with separate modules for core functionality, data models, and CLI
- Add .gitignore to exclude Python bytecode files
- Update README.md to reflect new project structure and describe each module's purpose
- Move main application logic from main.py to grokparty/app.py
- Split conversation loop and CLI handling into dedicated modules
- Improve code organization for better maintainability and scalability
