+++ .gitignore
@@ -1,0 +1,9 @@
+grokparty/models/__pycache__/conversation.cpython-313.pyc
+grokparty/models/__pycache__/__init__.cpython-313.pyc
+grokparty/models/__pycache__/character.cpython-313.pyc
+grokparty/core/__pycache__/grok_api.cpython-313.pyc
+grokparty/core/__pycache__/__init__.cpython-313.pyc
+grokparty/cli/__pycache__/cli.cpython-313.pyc
+grokparty/cli/__pycache__/__init__.cpython-313.pyc
+grokparty/__pycache__/app.cpython-313.pyc
+grokparty/__pycache__/__init__.cpython-313.pyc
