+++ grokparty/__init__.py
@@ -1,0 +1,4 @@
+"""
+GrokParty - Terminal-based AI Character Conversations
+A Python application that creates conversations between AI characters using Grok models.
+"""
