+++ grokparty/cli/__init__.py
@@ -1,0 +1,3 @@
+"""
+CLI module for GrokParty
+"""
