+++ grokparty/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Core module for GrokParty
+"""
