+++ grokparty/models/__init__.py
@@ -1,0 +1,3 @@
+"""
+Models module for GrokParty
+"""
